refactor: prepare for full_committer_flow benchmark
