+++ crates/committer_cli/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod block_hash;
+pub mod committer_runner;
 pub mod filled_tree_output;
 pub mod parse_input;
 pub mod tests;
